[*] Move wrapper code for command buffers and command pools into subfolder
